Add class DatasetMetaData and use it in ZarrDataset
